implements LIMIT paging for Cassandra Repository

+++ jkniv/jkniv-whinstone/jkniv-whinstone-cassandra/src/test/resources/database/db-create.sql
@@ -6,7 +6,7 @@ CREATE TABLE COLOR
 ( 
   NAME text,
   PRIORITY text,
-  CODE SMALLINT, 
+  CODE integer, 
  ,PRIMARY KEY( NAME, PRIORITY )
 );
